refactor(ros2agnocast): extract shared A2R bridge utilities into bridge_utils.py
Move load_msg_class and the dummy-subscription QoS/creation logic out of
_bridged_ros2cli.py into a new bridge_utils.py module so they can be shared
across the codebase.

- Add bridge_utils.py with load_msg_class, dummy_sub_qos, create_dummy_subscription
- Remove duplicated QoSProfile construction from _bridged_ros2cli.py and
  _a2r_bridge_activator.py; both now call create_dummy_subscription
- Fix _a2r_bridge_activator.py: was incorrectly importing load_msg_class from
  _bridged_ros2cli; now imports from bridge_utils
